Fix indentation error. Use __main__ idiom.
When initially trying the `example.py` file, it did not work. There was
an indentation error in the `for` block. I fixed that error by adding
the missing indentation.

I also placed the example code into an `if __name__ ...` block. This is
a common idiom used in Python to only run certain code when the file is
called directly. Code inside this block will not run when you use
`import discordflags`. This means it is a good place to include the
actual user interacting part of the code.

You can read more about the `if __name__ ...` main idiom, if you like,
at this very detailed StackOverflow answer:

  - https://stackoverflow.com/questions/419163/what-does-if-name-main-do