Add .gitignore
This file is used by `git` to determine which files should *not* be
included in the GitHub repository. Certain files are not meant to be
version controlled because they can cause issues, or contain private
information.

This `.gitignore` file is a copy of a template file provided by GitHub
for common files to ignore for Python projects. It is the same one I use
in my own project repos. The original file can be found at:

  - https://github.com/github/gitignore/blob/master/Python.gitignore