Place `flags.py` into `discordflags` module
This commit creates the actual `discordflags` module, and places the
existing `flags.py` into it. By storing a collection of Python source
files together, along with an `__init__.py` file (that can be empty),
you are telling Python that is a single module.

With that module, you can use statements like `import discordflags` in
your programs to use the code from the module. The `__init__.py` file
can be left blank, or module import, or module setup code can be added
in.

In this case, I chose to add `from .flags import flagsfromint` to the
module init file. This statement will import the `flagsfromint` function
defined in `discordflags/flags.py` whenever the `discordflags` module is
imported. If this `from` statement was not included, you would need to
use a statement like `import discordflags.flags` in the `example.py`

The comment block at the top of `discordflags/__init__.py` is used to
hold some basic module information, copied off of the GitHub repo. It is
not necessary, but can be helpful to include.