Update `example.py` to use new module
Update of the example code to utilize the new module created in the last
commit. The core change is in the `import` statement. By using `import
discordflags`, we tell Python to import the module located in the
`discordflags` directory. When importing the module, Python
automatically runs `discordflags/__init__.py` which is what brings in
the `flagsfromint` function.

The `.. as df` on the end of the `import` statement is basically
creating an alias, or shorter name, for `discordflags`. This means that
in the loop below, we will call `df.flagsfromint()` instead of
`discordflags.flagsfromint`. The inclusion of this `as df` part is
optional, but allows for less typing with  long module names.