Exit example.py gracefully
When exiting the `examply.py` program, there would be a few lines of
error output from Python. It looks a little bit cleaner if we show a
proper 'exit' so the user does not misinterpret that as an error.

By adding this `try` block we can `try` to do whatever is written
inside. This case, is the `input` statement. If there is either a
`KeyboardInterrupt` (Control-C) or `EOFError` (Control-D), we capture it
in the `except` block and display the error message before exiting. This
will only catch the two specific exceptions we listed. This is good
because other errors will still be displayed.